Allow building specific product in pipeline
Updated the pipeline to support building a specific product using the
BUILD_PRODUCT variable. When BUILD_PRODUCT is set only the specified product
will be built, and integration tests that are not relevant to that product
will be skipped. If it is not set, the pipeline will run as normal.

Closes GDO-256