Allow Docker users to define policy database creds
In PingDataGovernance-PAP 8.1, users have more control over the
credentials used to access the policy database (H2 file). However, due
to an oversight, the policy database admin credentials are unable to be
set via environment variables and instead must be specified as command
line arguments in the setup tool. This change modifies the setup command
in the Ping DevOps hook to do that and therefore allow Docker image
consumers to set the database admin credentials.

JiraIssue: DS-42383

Product: broker