Add variables for certificates
Updated the pingdataconsole, pingdelegator, and all pingdata images
with new variables that can be used to specify keystore, truststore,
and certificate files.

For pingdataconsole and pingdata images, use the KEYSTORE_FILE and
KEYSTORE_PIN_FILE variables to supply paths to a keystore and a
keystore password file to be used by the product. When these variables
are not specified, the default keystore paths will be used if they
exist. In pingdata these defaults are
${SERVER_ROOT_DIR}/config/keystore[.p12] and
${SERVER_ROOT_DIR}/config/keystore.pin. In pingdataconsole the defaults
are ${STAGING_DIR}/keystore and ${SECRETS_DIR}/.keystore.pin.

For pingdata images, use the TRUSTSTORE_FILE and TRUSTSTORE_PIN_FILE
variables to supply paths to a truststore and truststore password file
for the product. The default values are
${SERVER_ROOT_DIR}/config/truststore[.p12] and
${SERVER_ROOT_DIR}/config/truststore.pin.

For pingdelegator, use the CERTIFICATE_FILE and CERTIFICATE_KEY_FILE
variables to supply paths for the nginx cert and key. The default
values are ${SECRETS_DIR}/nginx.crt and ${SECRETS_DIR}/nginx.key.

This commit also fixes an issue where the container would not report
an error when invalid certificate, encryption, or JVM options were
specified in pingdata images, and an issue where PKCS12 keystores
and truststores were not handled correctly in 20-restart-sequence.sh
in pingdirectory.

Closes GDO-328