Update get_value to retrieve value from file
Updated the get_value function in pingcommon.lib.sh to support retrieving a
variable value from a file when that variable is set. For example, when
getting a value for a variable named PASSWORD, if PASSWORD isn't set, then
get_value can check if the PASSWORD_FILE variable is set. If it is, get_value
will read the value from that file.

get_value will not use this behavior by default. Pass in true as the second
argument to get_value to check the file when necessary.

Closes GDO-439