GDO-310: Print Warnings When Deprecated Variables are Used
Update the 04-check-variables.sh hook with warnings whenever a
deprecated variable has a set value. The warnings indicate that these
variables may be removed in a future release.

The /opt/staging/deprecated-variables.json file contains the list of
deprecated variables that are checked. The variables can be marked as
sensitive (such as for password variables) and can include a message
specific to that individual variable.

Closes GDO-310