Update PingDirectory certificate logic
This commit fixes an error that would occur on restart for pingdirectory
when using either a profile with no setup-arguments.txt file and no
keystore, or when using a profile with the
--generateSelfSignedCertificate argument in setup-arguments.txt. The
previous logic would copy the existing keystore and truststore files
into the server profile, which causes an exception when setting up the
server with a self-signed certificate.

Two main changes were made to fix this. First, if a profile has no
setup-arguments.txt file and one is generated by a hook, then that
generated file will be deleted after manage-profile completes. This
prevents the file sticking around between restarts. Second, if the
--generateSelfSignedCertificate argument is being used, then the
keystore and truststore files will not be copied into the profile
on restart, to prevent the exception during setup. Since manage-profile
replace-profile already maintains the default keystore and truststore
files, this will not cause a certificate change during restart.

Closes GDO-444