Run remove-defunct-server when volume is lost
Updated the PingDirectory and PingDataSync topology logic to run the
remove-defunct-server tool on restart when necessary. When the local
server's view of the topology does not match the seed server's view,
remove-defunct-server will be used to remove the local server and
re-add it to the topology. This situation is most likely to occur
when a container loses its persistent volume and is restarted.

Note: due to a bug in PingDirectory that was not fixed until version
8.2.0.0-EA (see DS-42438), this fix will only work on 8.2.0.0-EA or
later versions. Earlier versions will see an error message but will
require manual intervention on the seed server to fix.

Closes GDO-513