Add support for setting up the PDG PAP in OpenID Connect mode
To set up the PAP in OIDC mode, the user must specify two environment
variables:

* PING_OIDC_CONFIGURATION_ENDPOINT
* PING_CLIENT_ID

In addition, users are encouraged to set the PING_EXTERNAL_BASE_URL
variable, which the PAP uses to construct URLs used outside of the
container. This should be used in preference to REST_API_HOSTNAME,
which has been marked as deprecated.