Update pingdatagovernancepap for 8.2.0.0-GA
This refactors the pingdatagovernancepap version check into a
reusable function, and adds logic for a new required argument in `setup`,
which requires an `--adminPort`. It adds documentation to the Dockerfile
on how to modify default backup values. Finally, it updates the liveness.sh check to
use the /healthcheck endpoint when the admin connector is listening.

Reviewer: Jacob Childress

JiraIssue: DS-43341

Product: broker