Include docs necessary for PingData Swagger UI
Fix the Swagger UI page not working for PingData docker images. This
page depends on certain files in the docs folder in the server root,
which was previously completely deleted during product staging.

This change applies to the pingdirectory, pingdirectoryproxy, and
pingdatagovernance images. Other PingData producs do not include
Swagger docs.

GDO-549