Clean staging directory on image startup
Updated all product images to clean the /opt/staging directory on
startup. Any files found in /opt/staging that are not included in the
base image itself will be cleaned. The CLEAN_STAGING_DIR variable
can be set to false to prevent this behavior if necessary.

Also removed some unnecessary hooks - 21-update-server-profile.sh
and 188-apply-configuration.sh. 21-update-server-profile.sh isn't
needed as the profile is already pulled on a restart, and
188-apply-configuration.sh is not used anywhere. These hooks are
still called, so they can still be implemented in a server profile
if necessary.

GDO-668