Only regenerate java.properties when necessary
Updated the PingData images to reduce the frequency that
dsjavaproperties is run on container restart. It will now only be run
when either the JVM has changed, or the JVM options passed to setup
have changed. Otherwise the call will be skipped.

The new REGENERATE_JAVA_PROPERTIES variable can be set to "true" to
force regenerating the java.properties file if necessary.

GDO-667