Custom lib/_start-server.sh for performance
Add a new custom lib/_start-server.sh script to pingdatacommon. This
version of the script skips the precheck for start-server as well as
the calls to test-java in lib/_script-util.sh, which avoids three
separate java calls before the server is started. This should improve
performance for server startup and restart.

When or if there are PingData changes to lib/_start-server.sh that
allow skipping the pre-check and test-java logic without having to
provide a full custom script, we should remove the custom version
from pingdatacommon.

GDO-764