Add JVM option cache for PingData
Add a default JVM option cache to speed up setup and restart for
PingData images. The cache specifies sets of JVM arguments that
the setup tool will assume to be supported, rather than making
separate Java calls to explicitly test them. The cache filename
corresponds to the specific JVM that the image is running.

BRASS-5