[pingauthorizepap] Add PING_USE_CONTAINER_MESSAGES
This adds an undocumented environment variable to the pingauthorizepap
image, which is used to tell the product to use container-specific
messages when printing to stdout.

It also adds the capability to mark Dockerfile lines as
undocumented, which prevents them from being included in the generated
documentation.

Undocumented environment variables should be preceded with a comment
starting with a hash and two forward slashes. For example:

    #// The next uncommented line will remain undocumented.